Update problem.txt
I'd like to make it clearer we want to begin the promise chain with `iterate(1)`.

I also think the bonus section was a bit confusing. If you simply swap `console.log` with `throwMyGod`, the error is swallowed. I think the intent was to show that using `done` throws the error in the global scope, bringing it to the user's attention.