updates directions for exercise 3 to include link to MDN site on Error objects for clarification; also changes the formatting of the 'rule of thumb' in 'An Important Rule' to italic because the prior formatting was rendering in a hard-to-read dark blue on some black terminal screens.
changes wording to edit to directions for exercise 3 from previous commit

Modifies formatting of link in 'Reject A Promise' to flow better with the text. Also changes the formatting of the 'rule of thumb' in 'An Important Rule' to italic because the prior formatting was rendering in a hard-to-read dark blue on some black terminal screens.

Moves the MDN link from body of the assignment instructions down to the
Hint section.